services/firestore-crud-api-assessment02/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-firestore==2.13.0
pydantic